\name{lgs.case.only}
\alias{lgs.case.only}

\title{Locus Genome Statistics}

\description{Calculation of Locus Genome Statistics (LGS) for various
  models.}

\usage{
lgs.case.only(adm, phi1, phi2 = phi1^2, pop = 1, dev = FALSE)
}

\arguments{
  \item{adm}{An object returned by \code{\link{admixture}}.}
  \item{phi1}{Increased risk associated with one risk allele.}
  \item{phi2}{Increased risk associated with two risk alleles.}
  \item{pop}{Risk population.}
  \item{dev}{Logical value indicating the development version should be
    run.}
}

%\details{}

%\value{}

%\references{}

\author{Randall Johnson}

%\examples{}

\keyword{inference}
